+++ app/javascript/packs/application.js
@@ -8,7 +8,14 @@ require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
 
+import '@fortawesome/fontawesome-free/css/all.css'
 import '../css'
+import '../script'
+import TurbolinksAdapter from 'vue-turbolinks'
+import Vue from 'vue/dist/vue.esm'
+Vue.use(TurbolinksAdapter)
+global.Vue = Vue;
+
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
