+++ app/javascript/packs/hello_vue.js
@@ -22,25 +22,6 @@
 // yarn add vue-turbolinks
 // Then uncomment the code block below:
 
-import TurbolinksAdapter from 'vue-turbolinks'
-import Vue from 'vue/dist/vue.esm'
-import CreateBoard from '../templete/createboard.vue'
 
-Vue.use(TurbolinksAdapter)
 
-document.addEventListener('turbolinks:load', () => {
-  // const app = new Vue({
-  //   render: h => h(App)
-  // }).$mount()
-  // document.body.appendChild(app.$el)
 
-  // console.log(app)
-  const createBoard = new Vue({
-    el: "#createBoard",
-    methods: {
-      boardView: function(event){
-        console.log(event)
-      }
-    }
-  })
-})
