+++ app/javascript/script/application.js
@@ -1,0 +1,1 @@
+import "./board_controller.js"
