+++ app/javascript/script/board_controller.js
@@ -1,0 +1,16 @@
+document.addEventListener('turbolinks:load', () => {
+    // const app = new Vue({
+    //   render: h => h(App)
+    // }).$mount()
+    // document.body.appendChild(app.$el)
+  
+    // console.log(app)
+    const createBoard = new Vue({
+      el: "#createBoard",
+      methods: {
+        boardView: function(event){
+          console.log(event)
+        }
+      }
+    })
+  })
