+++ app/javascript/script/index.js
@@ -1,0 +1,1 @@
+import './application.js'
